requirements.txt
Django==3.2.16
numpy==1.26.4
Pillow==10.3.0
torch==2.3.1
ultralytics==8.2.31
opencv-python==4.10.0.82

